refactor(FavoriteSerice): check authorization
- 사용자의 즐겨찾기를 안전하게 보호하기 위해 본인의 즐겨찾기가 아니면 삭제할 수 없도록 기능 추가

+++ src/main/java/nextstep/subway/favorite/application/FavoriteService.java
@@ -2,19 +2,17 @@ package nextstep.subway.favorite.application;
 
 import nextstep.subway.auth.domain.LoginMember;
 import nextstep.subway.favorite.application.exceptions.FavoriteEntityNotFoundException;
+import nextstep.subway.favorite.application.exceptions.NotMyFavoriteException;
 import nextstep.subway.favorite.domain.Favorite;
 import nextstep.subway.favorite.domain.FavoriteRepository;
 import nextstep.subway.favorite.domain.adapters.SafeStationForFavoriteAdapter;
 import nextstep.subway.favorite.domain.excpetions.FavoriteCreationException;
 import nextstep.subway.favorite.ui.dto.FavoriteRequest;
 import nextstep.subway.favorite.ui.dto.FavoriteResponse;
-import nextstep.subway.favorite.ui.dto.StationInFavoriteResponse;
-import org.hibernate.exception.ConstraintViolationException;
 import org.springframework.dao.DataIntegrityViolationException;
-import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
-import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -28,12 +26,11 @@ public class FavoriteService {
         this.safeStationAdapter = safeStationAdapter;
     }
 
-    public void deleteFavorite(final Long deleteTarget) {
-        try {
-            favoriteRepository.deleteById(deleteTarget);
-        } catch (EmptyResultDataAccessException e) {
-            throw new FavoriteEntityNotFoundException("존재하지 않는 즐겨찾기입니다.");
-        }
+    public void deleteFavorite(LoginMember loginMember, Long deleteTarget) {
+        Favorite favorite = this.getFavoriteById(deleteTarget);
+        validateIsOwnerOfFavorite(loginMember, favorite);
+
+        favoriteRepository.deleteById(deleteTarget);
     }
 
     public Long saveFavorite(LoginMember loginMember, FavoriteRequest favoriteRequest) {
@@ -49,12 +46,13 @@ public class FavoriteService {
         }
     }
 
-    private void validateStationsExists(final FavoriteRequest favoriteRequest) {
-        if(!safeStationAdapter.isAllExists(favoriteRequest.getSource(), favoriteRequest.getTarget())) {
-            throw new FavoriteCreationException("존재하지 않는 역으로 즐겨찾기를 생성할 수 없습니다.");
-        }
+    @Transactional(readOnly = true)
+    public Favorite getFavoriteById(Long id) {
+        return favoriteRepository.findById(id)
+                .orElseThrow(() -> new FavoriteEntityNotFoundException("존재하지 않는 즐겨찾기 입니다."));
     }
 
+    @Transactional(readOnly = true)
     public List<FavoriteResponse> getFavorites(LoginMember loginMember) {
         List<Favorite> favorites = favoriteRepository.findAllByMemberId(loginMember.getId());
 
@@ -63,5 +61,17 @@ public class FavoriteService {
                 safeStationAdapter.getSafeStationInFavorite(it.getSourceId()),
                 safeStationAdapter.getSafeStationInFavorite(it.getTargetId()))
         ).collect(Collectors.toList());
+    }
+
+    private void validateIsOwnerOfFavorite(final LoginMember loginMember, final Favorite favorite) {
+        if (!favorite.isOwner(loginMember.getId())) {
+            throw new NotMyFavoriteException("본인의 즐겨찾기만 삭제 가능합니다.");
+        }
+    }
+
+    private void validateStationsExists(final FavoriteRequest favoriteRequest) {
+        if(!safeStationAdapter.isAllExists(favoriteRequest.getSource(), favoriteRequest.getTarget())) {
+            throw new FavoriteCreationException("존재하지 않는 역으로 즐겨찾기를 생성할 수 없습니다.");
+        }
     }
 }
